MM-10666: ignore unknown profiles in `sortAndInjectProfiles` (#505) (#509)
* test unknown user id in current channel

Verify that the `getProfilesInCurrentChannel` selector handles a user id
in the current channel for which there is no profile (yet).

* MM-10666: ignore unknown profiles in `sortAndInjectProfiles` (#505)

The `profileSet` might reference a user id that is not yet known to the
`profiles` -- in that case, filter out the profile and wait for the
reducer to update with same.