[Documentation] simplified the "Promise" example (#784)
#### Summary
If someone were to use promise, it would be clearer to write their code this way, doing so they would take advantage of the promises.

#### Ticket Link
There is no ticket for that tell me if I need to create one (in which case I will strongly reconsider contributing this piece of doc since it would mean creating an account on a third party site and all ...)

#### Test Information
This PR was not tested at all. The mentionned code was tested in a console on chrome